requirements.txt
# gremlinsAI Backend Dependencies
# Production-ready versions for Phase 1 & Phase 2

# FastAPI and Server
fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.20.0,<1.0.0

# LangChain and LangGraph - Core AI Framework
langchain>=0.1.0,<1.0.0
langgraph>=0.0.40,<1.0.0
langchain-community>=0.0.30,<1.0.0
langchain-core>=0.1.40,<1.0.0

# Search Tools
duckduckgo-search>=6.0.0,<7.0.0

# Database (Phase 2)
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.13.0,<2.0.0
aiosqlite>=0.19.0,<1.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.0.0,<3.0.0

# OAuth2 Authentication
PyJWT>=2.8.0,<3.0.0
cryptography>=41.0.0,<42.0.0
httpx>=0.25.0,<1.0.0

# Phase 3: Multi-Agent Architecture
crewai>=0.28.0,<1.0.0
crewai-tools>=0.1.0,<1.0.0
langchain-openai>=0.1.0,<1.0.0

# Phase 4: Data Infrastructure Overhaul - Weaviate Vector Database
weaviate-client>=4.4.0,<5.0.0
sentence-transformers>=2.2.0,<3.0.0
langchain-weaviate>=0.0.1,<1.0.0
pypdf>=4.0.0,<5.0.0
python-multipart>=0.0.6,<1.0.0

# CLIP embeddings for multimodal search
clip-by-openai>=1.0.0,<2.0.0

# Phase 5: Orchestration & Scalability
celery>=5.3.0,<6.0.0
redis>=5.0.0,<6.0.0
kombu>=5.3.0,<6.0.0
billiard>=4.2.0,<5.0.0

# Phase 6: API Modernization & Real-time Communication
strawberry-graphql[fastapi]>=0.215.0,<1.0.0
websockets>=12.0,<13.0

# Phase 7: Multi-Modal Revolution
# Audio processing
openai-whisper>=20231117
pydub>=0.25.1,<1.0.0
pyttsx3>=2.90,<3.0.0
librosa>=0.10.0,<1.0.0

# Video processing
ffmpeg-python>=0.2.0,<1.0.0
opencv-python>=4.8.0,<5.0.0

# Image processing
Pillow>=10.0.0,<11.0.0
torch>=2.0.0,<3.0.0
torchvision>=0.15.0,<1.0.0

# Computer vision and OCR (optional)
# pytesseract>=0.3.10,<1.0.0

# Phase 8: Developer Enablement & Documentation
jinja2>=3.1.0,<4.0.0
click>=8.1.0,<9.0.0
rich>=13.0.0,<14.0.0
httpx>=0.25.0,<1.0.0

# Local LLM Support
langchain-ollama>=0.1.0,<1.0.0
langchain-huggingface>=0.0.3,<1.0.0
transformers>=4.35.0,<5.0.0
accelerate>=0.24.0,<1.0.0
